stupid type comparison bug.   ((int) -1 > (size_t) 1234) returns true.  In this case, content_length is an int but maxObjectSize is a size_t.  Objects without content length were not being cached because storeCheckCachable said they were too big.

+++ src/store.c
@@ -546,7 +546,7 @@ storeCheckCachable(StoreEntry * e)
     } else if (e->mem_obj->inmem_hi > Config.Store.maxObjectSize) {
 	debug(20, 2) ("storeCheckCachable: NO: too big\n");
 	store_check_cachable_hist.no.too_big++;
-    } else if (e->mem_obj->reply->content_length > Config.Store.maxObjectSize) {
+    } else if (e->mem_obj->reply->content_length > (int) Config.Store.maxObjectSize) {
 	debug(20, 2) ("storeCheckCachable: NO: too big\n");
 	store_check_cachable_hist.no.too_big++;
     } else if (EBIT_TEST(e->flags, KEY_PRIVATE)) {
